Updates rotation and bullet behavior
Sorry for the delay, if you remind me, I will discount the late pentaly by four days. Part of the problem you were having was relying too much on the rotation_degrees (which is managed by the physics engine) rather than calculating it yourself. I tried to make everything based on a global variable in the Game script, so you could rotate at a different angle than 90 if you choose.